+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * Added relics: Spicy Sausage, Treasure Map, Bottled Yo-Yo
 * Fix: Water orb shouldn't be affected by focus
 * Fix: Necrotic Wound should not play at end of turn when you have Runic Pyramid
+* Fix: Purge shouldn't remove Necronomicurse, Ascender's Bane, and other souldbound cards (interacts with StsLib)
 
 ### v1.4.1
 * In multi monster fights, the left monsters don't apply vulnerable.